backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.4.0
numpy==2.2.6
librosa==0.10.2
soundfile==0.12.1
google-generativeai==0.5.4
werkzeug==3.0.3
gunicorn==22.0.0
